Refactoring lists
